refactor: renamed files and folders in pascal case
