+++ components/auth/ForgotPassword/ForgotPassword.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import style from './ForgotPassword.module.scss'
-import { MainButton } from '@/components/common/Buttons/buttons'
-import { Loading } from '@/components/common/loaders/Loading'
+import { MainButton } from '@/components/common/Buttons/Buttons'
+import { Loading } from '@/components/common/Loaders/Loading'
 import ReCAPTCHA from 'react-google-recaptcha'
 import { IForgotPasswordProps } from './ForgotPasswordTypes'
 import ForgotPasswordInput from './ForgotPasswordInput'
