+++ components/auth/Login/LoginFormContainer.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren, useEffect } from 'react'
 import { useRouter } from 'next/router'
-import { useLoginMutation } from '@/assets/api/auth/authApi'
-import { ServerLoginResponse } from '@/assets/api/auth/authTypes'
+import { useLoginMutation } from '@/assets/api/Auth/AuthApi'
+import { ServerLoginResponse } from '@/assets/api/Auth/AuthTypes'
 import LoginForm from './LoginForm'
 import style from './LoginForm.module.scss'
 import { useState } from 'react'
