+++ components/auth/Registration/RegistrationFormContainer.tsx
@@ -1,8 +1,8 @@
 import RegistrationForm from "./RegistrationForm"
-import { ErrorMessagerType, FormValuesType, PrintModalType } from "./type"
-import { useRegistrationMutation } from "@/assets/api/auth/authApi"
+import { ErrorMessagerType, FormValuesType, PrintModalType } from "./Type"
+import { useRegistrationMutation } from "@/assets/api/Auth/AuthApi"
 import { useState } from "react"
-import { Modal } from "@/components/common/Modal/modal"
+import { Modal } from "@/components/common/Modal/Modal"
