+++ components/auth/failed/FailedAuth.tsx
@@ -2,9 +2,9 @@ import Image from 'next/image'
 import img from '../../../public/img/time-management.png'
 import { PageWrapper } from '@/components/PageWrapper/PageWrapper'
 import style from './FailedAuth.module.scss'
-import commonStyle from '../../../styles/common.module.scss'
-import { MainButton } from '@/components/common/Buttons/buttons'
-import { useRegistrationEmailResendMutation } from '@/assets/api/auth/authApi'
+import commonStyle from '../../../styles/Common.module.scss'
+import { MainButton } from '@/components/common/Buttons/Buttons'
+import { useRegistrationEmailResendMutation } from '@/assets/api/Auth/AuthApi'
 import { useEffect, useState } from 'react'
 
 const FailedAuth = () => {
