+++ components/auth/success/SuccessAuth.tsx
@@ -1,11 +1,11 @@
 import Image from 'next/image'
 import img from '../../../public/img/wall-success.png'
 import style from './SuccessAuth.module.scss'
-import commonStyle from '../../../styles/common.module.scss'
+import commonStyle from '../../../styles/Common.module.scss'
 import { PageWrapper } from '@/components/PageWrapper/PageWrapper'
 import { useRouter } from 'next/router'
-import { MainButton } from '@/components/common/Buttons/buttons'
-import { authRouts } from '@/components/common/Auth/authRouts'
+import { MainButton } from '@/components/common/Buttons/Buttons'
+import { authRouts } from '@/components/common/Auth/AuthRoutes'
 
 const SuccessAuth = () => {
   const router = useRouter()
