+++ components/common/Auth/LogoGroup.module.scss
