+++ components/common/loaders/Loading.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import LoadingIcon from '@/public/icons/loading.svg'
-import style from './loading.module.scss'
+import style from './Loading.module.scss'
 
 export const Loading = () => {
   return (
