+++ pages/auth/log-out/index.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
 import { getLayout } from '@/components/Layout/Layout'
-import { useLogoutMutation } from '@/assets/api/auth/authApi'
+import { useLogoutMutation } from '@/assets/api/Auth/AuthApi'
 import { PageWrapper } from '@/components/PageWrapper/PageWrapper'
-import { Loading } from '@/components/common/loaders/Loading'
+import { Loading } from '@/components/common/Loaders/Loading'
 
 const LogOut = () => {
     const router = useRouter()
