+++ pages/index.tsx
@@ -3,9 +3,9 @@ import { NextPageWithLayout } from './_app'
 import { getLayout } from '@/components/Layout/Layout'
 import Login from './auth/login'
 import { useRouter } from 'next/router'
-import { useMeQuery } from '@/assets/api/auth/authApi'
+import { useMeQuery } from '@/assets/api/Auth/AuthApi'
 import { useEffect } from 'react'
-import { Loading } from '@/components/common/loaders/Loading'
+import { Loading } from '@/components/common/Loaders/Loading'
 
 const Home: NextPageWithLayout = () => {
   const router = useRouter()
